+++ Tutorial/Source/Streets.cpp
@@ -1,5 +1,5 @@
 #include <algorithm>
 
-float randRange(float min, float max) {
-    return min + static_cast <float> (rand()) / (static_cast <float> (RAND_MAX / (max - min)));
+double randRange(double min, double max) {
+    return min + static_cast <double> (rand()) / (static_cast <double> (RAND_MAX / (max - min)));
 }
